Limit batch count in UntexturedObjectsGLBufferRange test.
Previous implementation was querying the maximum size of a UBO and then attempting to make a system memory allocation of that size. An implementation could theoretically support up to a 4GiB UBO. Such an allocation attempt would obviously fail on a 32-bit system. Implicitly limit the size of the attempted allocation by clamping the batch size to the minimum of the object count (i.e., one big batch), or 64^3, whichever is smaller.